requirements/test.txt
django==1.5.4
djangorestframework==2.3.8
django-rest-swagger==0.1.8
Markdown==2.3.1
django-cms==2.4.2
Pillow==2.0.0
django-cors-headers==0.11
boto==2.13.3
django-storages==1.1.8
#git+git://github.com/mpaolini/django-versioned-storages@v0.1
git+git://github.com/gg4u/django-versioned-storages@v0.1
# needed by django-storages boto backend
pytz==2013d
python-dateutil==2.1
django-json-field==0.5.5
django-filter==0.7
djangocms-text-ckeditor==1.0.10
# for development
fabric==1.8
nodeenv==0.7.0
